requirements.txt
# ── Core Data & Finance ─────────────────────────────────────
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0

# ── Machine Learning ────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0
hmmlearn>=0.3.0
shap>=0.44.0

# ── Deep Learning ───────────────────────────────────────────
torch>=2.0.0
transformers>=4.35.0          # FinBERT via HuggingFace

# ── Reinforcement Learning ───────────────────────────────────
stable-baselines3>=2.2.0
gymnasium>=0.29.0

# ── Optimization ────────────────────────────────────────────
cvxpy>=1.4.0
PyPortfolioOpt>=1.5.5

# ── Macro Data ──────────────────────────────────────────────
requests>=2.31.0
pandas-datareader>=0.10.0

# ── Visualization ───────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# ── Utilities ───────────────────────────────────────────────
tqdm>=4.66.0
pyarrow>=13.0.0               # Parquet cache
joblib>=1.3.0
